+++ pytest_subtests.py
@@ -97,6 +97,27 @@ def pytest_configure(config):
     TestCaseFunction.addSubTest = _addSubTest
     TestCaseFunction.failfast = False
 
+    # Hack (#86): the terminal does not know about the "subtests"
+    # status, so it will by default turn the output to yellow.
+    # This forcibly adds the new 'subtests' status.
+    import _pytest.terminal
+
+    new_types = tuple(
+        f"subtests {outcome}" for outcome in ("passed", "failed", "skipped")
+    )
+    # We need to check if we are not re-adding because we run our own tests
+    # with pytester in-process mode, so this will be called multiple times.
+    if new_types[0] not in _pytest.terminal.KNOWN_TYPES:
+        _pytest.terminal.KNOWN_TYPES = _pytest.terminal.KNOWN_TYPES + new_types
+
+    _pytest.terminal._color_for_type.update(
+        {
+            f"subtests {outcome}": _pytest.terminal._color_for_type[outcome]
+            for outcome in ("passed", "failed", "skipped")
+            if outcome in _pytest.terminal._color_for_type
+        }
+    )
+
 
 def pytest_unconfigure():
     if hasattr(TestCaseFunction, "addSubTest"):